requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-community
openai
pinecone
pydantic
python-dotenv
unstructured
pypdf
